refactor: rename api/server to services for clearer layer separation
